backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-filter==24.3
Pillow==10.4.0
psycopg[binary,pool]==3.2.1
python-dotenv==1.0.1
django-cors-headers==4.4.0
reportlab==4.2.2